backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pydantic==2.12.5
pydantic-settings==2.12.0
openai==1.57.2
pinecone-client==5.0.1
openpyxl==3.1.5
requests==2.32.3
aiohttp==3.13.2
python-dotenv==1.0.1
tiktoken==0.12.0